// frontend/food-ordering-app/src/model/enum.ts
export enum DeliveryStatus {
  WAITING_FOR_KITCHEN = "WAITING_FOR_KITCHEN",
  AT_DELIVERY = "AT_DELIVERY",
  DELIVERED = "DELIVERED",
}

export enum KitchenTicketStatus {
  PREPARING = "PREPARING",
  READY_FOR_DELIVERY = "READY_FOR_DELIVERY",
  REJECTED = "REJECTED",
  FINISHED = "FINISHED"
}
export enum OrderStatus {
  PENDING = "PENDING",
  PAID = "PAID",
  APPROVED = "APPROVED",
  COMPLETED = "COMPLETED",
  CANCELLING = "CANCELLING",
  CANCELLED = "CANCELLED"
}